+++ lib/src/name_changer_methods/change_android_name.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:dcli/dcli.dart';
+import 'package:path/path.dart';
 
 /// Changes Android name of the project.
 void changeAndroidName({