+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.0
-
 - Initial version.
 
 ## 1.0.1
-- Added executable to use it from everywhere
+- Supports clone the target repo with specific branch.
+- 
+## 1.0.2
+- fix: cannot clone the repo with target branch